agent-fastpy/requirements.txt
dashscope>=1.10.0
python-dotenv>=1.0.0
requests>=2.31.0
typing>=3.7.4.3
aiohttp>=3.8.0
Pillow>=10.0.0
asyncio>=3.4.3
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
sqlalchemy>=2.0.0